Revert "Merge pull request #4 from Jellymaxd/xudongbranch"
This reverts commit cad6623e37bd4f1c6c651ea5617eaa4f5393d95a, reversing
changes made to 82f9696ae61b6cce182a3ebe18dca40b6a0b7265.
