+++ Projects/app/google-services.json
@@ -46,4 +46,3 @@
   ],
   "configuration_version": "1"
 }
-
