+++ Projects/app/src/main/AndroidManifest.xml
@@ -30,9 +30,20 @@
             android:label="@string/title_activity_login"
             android:parentActivityName=".SplashActivity">
             <meta-data
-               android:name="android.support.PARENT_ACTIVITY"
+                android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.example.cycleasy.SplashActivity" />
         </activity>
+
+        <activity android:name=".SignupActivity"></activity>
+        <activity
+            android:name=".ui.login.LoginActivity"
+            android:label="@string/title_activity_login"
+            android:parentActivityName=".SplashActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value="com.example.cycleasy.SplashActivity" />
+        </activity>
+
         <activity android:name=".ContentActivity" />
         <activity
             android:name=".SplashActivity"
