+++ Projects/app/src/main/java/com/example/cycleasy/MainActivity.java
@@ -17,48 +17,12 @@ import com.google.android.material.bottomnavigation.BottomNavigationItemView;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 
-public class MainActivity extends AppCompatActivity
+public class MainActivity extends AppCompatActivity {
 
-    implements BottomNavigationView.OnNavigationItemSelectedListener  {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
 
-        @Override
-        protected void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            setContentView(R.layout.activity_main);
-
-            BottomNavigationView navigation = findViewById(R.id.bottom_nav);
-            navigation.setOnNavigationItemSelectedListener(this);
-            loadfragment(new RouteFragment());
-
-
-        }
-        private boolean loadfragment(Fragment fragment){
-            if (fragment!=null){
-                getSupportFragmentManager()
-                        .beginTransaction()
-                        .replace(R.id.fragment_container,fragment)
-                        .commit();
-                return true;
-            }
-            return false;
-        }
-        @Override
-        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-            Fragment fragment=null;
-            switch (item.getItemId()){
-                case R.id.menu_exercise:
-                    fragment=new ExerciseFragment();
-                    break;
-                case R.id.menu_route:
-                    fragment=new RouteFragment();
-                    break;
-                case R.id.menu_racks:
-                    fragment=new RacksFragment();
-                    break;
-                case R.id.menu_me:
-                    fragment=new MeFragment();
-                    break;
-            }
-            return loadfragment(fragment);
-        }
     }
+}
