+++ Projects/app/src/main/res/values/styles.xml
@@ -20,7 +20,7 @@
 
     <style name="SplashTheme" parent="Theme.AppCompat.NoActionBar">
         <item name="android:windowContentOverlay">@null</item>
-        <item name="android:windowBackground">@drawable/appiconf</item>
+        <item name="android:windowBackground">@drawable/background_splash</item>
         <item name="android:windowFullscreen">true</item>
     </style>
